The final version
Only the interface design and game story design is not complete, but it
has little to do with the code. I deleted the game story editor to make
this project more like a game.In the future, I will separate out the
logical framework of the game and this game will serve as an example of
the logical framework.